requirements.txt
lxml==4.9.3
Pillow==10.1.0
python-pptx==0.6.23
XlsxWriter==3.1.9
